Building a DSPy pipeline for the generation to optimize the prompts.
This commit contains the fetching of chunks from the existing filtered chunks.
It is tested.

Also with it, comes some prompts used to generate the code with Gemini.
There is the prompt to generate the plan,
the prompt that integrates the plan (modified to add clarity on certain points)
and the prompt that gives further instruction to execute step 1.2