feat(services) : First robust version of the Excel parser
This version handles title, summary of tables and rows.
I had to rewrite the excel_processor almost entirely to build a better logic.
Now the processor handles file level operations and leaves sheet level operations to the sheet classes.

The qdrant_service is updated with a new collection to allow for querying the excel document first and then going into the table rows.